some text changes
This is the next part of the major refactoring of the moves format, moving most things from Move_Basic to MoveEntry.

+++ src/main/java/pokecube/api/entity/pokemob/IPokemob.java
@@ -31,12 +31,12 @@ import pokecube.api.data.pokedex.DefaultFormeHolder;
 import pokecube.api.data.spawns.SpawnRule;
 import pokecube.api.entity.pokemob.ai.CombatStates;
 import pokecube.api.entity.pokemob.ai.LogicStates;
-import pokecube.api.moves.IMoveConstants;
+import pokecube.api.moves.Battle;
+import pokecube.api.moves.utils.IMoveConstants;
 import pokecube.api.utils.PokeType;
 import pokecube.core.PokecubeCore;
 import pokecube.core.PokecubeItems;
 import pokecube.core.database.Database;
-import pokecube.core.moves.Battle;
 import thut.api.ModelHolder;
 import thut.api.entity.ICopyMob;
 import thut.api.entity.IHungrymob;
@@ -416,7 +416,7 @@ public interface IPokemob extends IHasMobAIStates, IHasMoves, ICanEvolve, IHasOw
      *
      * @return the status
      */
-    byte getStatus();
+    int getStatus();
 
     /**
      * The timer for SLP. When reach 0, the mob wakes up.
